delete directory only if exists

+++ job-manager/src/manager.rs
@@ -203,9 +203,12 @@ impl JobManagerState {
                 .await?;
             self.task_repository.delete_by_id(&ot.id).await?;
         }
-        if let Err(e) = retry_fs::remove_dir_all(job.root_dir).await {
-            error!("could not delete directory {e}");
+        if job.root_dir.exists() {
+            if let Err(e) = retry_fs::remove_dir_all(job.root_dir).await {
+                error!("could not delete directory {e}");
+            }
         }
+
         self.job_repository.delete_by_id(&job.id).await?;
         Ok(())
     }
